Refactor parsing config
