+++ cmd/executor.go
@@ -42,7 +42,7 @@ func BuildDefaultExecutor(config *cmdtypes.Config) cli.Executor {
 	rootCmd.AddCommand(
 		VersionCmd(),
 		initcmd.InitCmd(config.GetInitConfig()),
-		parsecmd.ParseCmd(config.GetParseConfig()),
+		parsecmd.ParseCmd(config),
 		snapshotcmd.ImportSnapshotCmd(config.GetSnapshotConfig()),
 		dbcmd.DbCmd(config.GetDbConfig()),
 		tokenlistcmd.ImportTokenListCmd(config),
