--- cmd/parse/config.go
@@ -1,33 +1,0 @@
-package parse
-
-import (
-	"fmt"
-	"os"
-
-	"github.com/spf13/cobra"
-
-	"github.com/forbole/soljuno/types"
-)
-
-// ReadConfig parses the configuration file for the executable having the give name using
-// the provided configuration parser
-func ReadConfig(cfg *Config) types.CobraCmdFunc {
-	return func(_ *cobra.Command, _ []string) error {
-		file := types.GetConfigFilePath()
-
-		// Make sure the path exists
-		if _, err := os.Stat(file); os.IsNotExist(err) {
-			return fmt.Errorf("config file does not exist. Make sure you have run the init command")
-		}
-
-		// Read the config
-		cfg, err := types.Read(file, cfg.GetConfigParser())
-		if err != nil {
-			return err
-		}
-
-		// Set the global configuration
-		types.Cfg = cfg
-		return nil
-	}
-}
