+++ cmd/parse/types.go
@@ -5,7 +5,6 @@ import (
 	"github.com/panjf2000/ants/v2"
 
 	"github.com/forbole/soljuno/db"
-	"github.com/forbole/soljuno/db/builder"
 	"github.com/forbole/soljuno/modules"
 	"github.com/forbole/soljuno/modules/registrar"
 	"github.com/forbole/soljuno/solana/client"
@@ -13,72 +12,11 @@ import (
 )
 
 // Config contains all the configuration for the "parse" command
-type Config struct {
-	registrar    registrar.Registrar
-	configParser types.ConfigParser
-	buildDb      db.Builder
-	logger       logging.Logger
-}
-
-// NewConfig allows to build a new Config instance
-func NewConfig() *Config {
-	return &Config{}
-}
-
-// WithRegistrar sets the modules registrar to be used
-func (config *Config) WithRegistrar(r registrar.Registrar) *Config {
-	config.registrar = r
-	return config
-}
-
-// GetRegistrar returns the modules registrar to be used
-func (config *Config) GetRegistrar() registrar.Registrar {
-	if config.registrar == nil {
-		return &registrar.EmptyRegistrar{}
-	}
-	return config.registrar
-}
-
-// WithConfigParser sets the configuration parser to be used
-func (config *Config) WithConfigParser(p types.ConfigParser) *Config {
-	config.configParser = p
-	return config
-}
-
-// GetConfigParser returns the configuration parser to be used
-func (config *Config) GetConfigParser() types.ConfigParser {
-	if config.configParser == nil {
-		return types.DefaultConfigParser
-	}
-	return config.configParser
-}
-
-// WithDBBuilder sets the database builder to be used
-func (config *Config) WithDBBuilder(b db.Builder) *Config {
-	config.buildDb = b
-	return config
-}
-
-// GetDBBuilder returns the database builder to be used
-func (config *Config) GetDBBuilder() db.Builder {
-	if config.buildDb == nil {
-		return builder.Builder
-	}
-	return config.buildDb
-}
-
-// WithLogger sets the logger to be used while parsing the data
-func (config *Config) WithLogger(logger logging.Logger) *Config {
-	config.logger = logger
-	return config
-}
-
-// GetLogger returns the logger to be used when parsing the data
-func (config *Config) GetLogger() logging.Logger {
-	if config.logger == nil {
-		return logging.DefaultLogger()
-	}
-	return config.logger
+type Config interface {
+	GetRegistrar() registrar.Registrar
+	GetConfigParser() types.ConfigParser
+	GetDBBuilder() db.Builder
+	GetLogger() logging.Logger
 }
 
 // --------------------------------------------------------------------------------------------------------------------
