+++ cmd/soljuno/main.go
@@ -4,19 +4,13 @@ import (
 	"os"
 
 	"github.com/forbole/soljuno/cmd"
-	"github.com/forbole/soljuno/cmd/parse"
-
-	"github.com/forbole/soljuno/modules/registrar"
 
 	cmdtypes "github.com/forbole/soljuno/cmd/types"
 )
 
 func main() {
 	// ParsingConfig the runner
-	config := cmdtypes.NewConfig("soljuno").
-		WithParseConfig(parse.NewConfig().
-			WithRegistrar(registrar.NewDefaultRegistrar()),
-		)
+	config := cmdtypes.NewConfig("soljuno")
 
 	// Run the commands and panic on any error
 	exec := cmd.BuildDefaultExecutor(config)
